+++ pymodbus/client/__init__.py
@@ -2,15 +2,11 @@
 
 import external classes, to make them easier to use:
 """
-from pymodbus.client.async_serial import AsyncModbusSerialClient
-from pymodbus.client.async_tcp import AsyncModbusTcpClient
-from pymodbus.client.async_tls import AsyncModbusTlsClient
-from pymodbus.client.async_udp import AsyncModbusUdpClient
+from pymodbus.client.serial import AsyncModbusSerialClient, ModbusSerialClient
+from pymodbus.client.tcp import AsyncModbusTcpClient, ModbusTcpClient
+from pymodbus.client.tls import AsyncModbusTlsClient, ModbusTlsClient
+from pymodbus.client.udp import AsyncModbusUdpClient, ModbusUdpClient
 from pymodbus.client.base import ModbusBaseClient
-from pymodbus.client.sync_serial import ModbusSerialClient
-from pymodbus.client.sync_tcp import ModbusTcpClient
-from pymodbus.client.sync_tls import ModbusTlsClient
-from pymodbus.client.sync_udp import ModbusUdpClient
 
 
 # ---------------------------------------------------------------------------#
