+++ pymodbus/client/tls.py
@@ -1,15 +1,15 @@
 """Modbus client async TLS communication."""
 import asyncio
 import logging
+import socket
 import ssl
 
-from pymodbus.client.async_tcp import AsyncModbusTcpClient
+from pymodbus.client.tcp import AsyncModbusTcpClient, ModbusTcpClient
 from pymodbus.client.base import ModbusClientProtocol
 from pymodbus.constants import Defaults
 from pymodbus.framer import ModbusFramer
 from pymodbus.framer.tls_framer import ModbusTlsFramer
 from pymodbus.transaction import FifoTransactionManager
-
 
 _logger = logging.getLogger(__name__)
 
@@ -84,13 +84,14 @@ class AsyncModbusTlsClient(AsyncModbusTcpClient):
         """Initialize Asyncio Modbus TLS Client."""
         super().__init__(host, port=port, framer=framer, **kwargs)
         self.sslctx = sslctx_provider(sslctx, certfile, keyfile, password)
-        self.certfile = certfile
-        self.keyfile = keyfile
-        self.password = password
-        self.server_hostname = server_hostname
+        self.params.sslctx = sslctx
+        self.params.certfile = certfile
+        self.params.keyfile = keyfile
+        self.params.password = password
+        self.params.server_hostname = server_hostname
 
         if not sslctx:
-            self.sslctx = ssl.create_default_context()
+            self.params.sslctx = ssl.create_default_context()
             # According to MODBUS/TCP Security Protocol Specification, it is
             # TLSv2 at least
             self.sslctx.options |= ssl.OP_NO_TLSv1_1
@@ -99,10 +100,6 @@ class AsyncModbusTlsClient(AsyncModbusTcpClient):
             self.sslctx.options |= ssl.OP_NO_SSLv2
         else:
             self.sslctx = sslctx
-        self.certfile = certfile
-        self.keyfile = keyfile
-        self.password = password
-        self.server_hostname = server_hostname
         AsyncModbusTcpClient.__init__(self, host, port=port, framer=framer, **kwargs)
 
     async def connect(self):
@@ -120,7 +117,7 @@ class AsyncModbusTlsClient(AsyncModbusTcpClient):
                 self.params.host,
                 self.params.port,
                 ssl=self.sslctx,
-                server_hostname=self.server_hostname,
+                server_hostname=self.params.server_hostname,
             )
         except Exception as exc:  # pylint: disable=broad-except
             txt = f"Failed to connect: {exc}"
@@ -137,3 +134,89 @@ class AsyncModbusTlsClient(AsyncModbusTcpClient):
         protocol.transaction = FifoTransactionManager(self)
         protocol.factory = self
         return protocol
+
+
+class ModbusTlsClient(ModbusTcpClient):
+    """**ModbusTlsClient**.
+
+    :param host: Host IP address or host name
+    :param port: (optional) Port used for communication.
+    :param framer: (optional) Framer class.
+    :param source_address: (optional) Source address of client,
+    :param sslctx: (optional) SSLContext to use for TLS
+    :param certfile: (optional) Cert file path for TLS server request
+    :param keyfile: (optional) Key file path for TLS server request
+    :param password: (optional) Password for for decrypting private key file
+    :param server_hostname: (optional) Bind certificate to host,
+    :param kwargs: (optional) Experimental parameters.
+
+    Example::
+
+        from pymodbus.client import ModbusTlsClient
+
+        async def run():
+            client = ModbusTlsClient("localhost")
+
+            client.connect()
+            ...
+            client.close()
+    """
+
+    def __init__(
+        self,
+        host: str,
+        port: int = Defaults.TlsPort,
+        framer: ModbusFramer = ModbusTlsFramer,
+        sslctx: str = None,
+        certfile: str = None,
+        keyfile: str = None,
+        password: str = None,
+        server_hostname: str = None,
+        **kwargs: any,
+    ):
+        """Initialize Modbus TLS Client."""
+        super().__init__(host, port=port, framer=framer, **kwargs)
+        self.sslctx = sslctx_provider(sslctx, certfile, keyfile, password)
+        self.params.sslctx = sslctx
+        self.params.certfile = certfile
+        self.params.keyfile = keyfile
+        self.params.password = password
+        self.params.server_hostname = server_hostname
+
+    @property
+    def connected(self):
+        """Connect internal."""
+        return self.connect()
+
+    def connect(self):
+        """Connect to the modbus tls server."""
+        if self.socket:
+            return True
+        try:
+            sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+            if self.params.source_address:
+                sock.bind(self.params.source_address)
+            self.socket = self.sslctx.wrap_socket(
+                sock, server_side=False, server_hostname=self.params.host
+            )
+            self.socket.settimeout(self.params.timeout)
+            self.socket.connect((self.params.host, self.params.port))
+        except socket.error as msg:
+            txt = (
+                f"Connection to ({self.params.host}, {self.params.port}) failed: {msg}"
+            )
+            _logger.error(txt)
+            self.close()
+        return self.socket is not None
+
+    def __str__(self):
+        """Build a string representation of the connection."""
+        return f"ModbusTlsClient({self.params.host}:{self.params.port})"
+
+    def __repr__(self):
+        """Return string representation."""
+        return (
+            f"<{self.__class__.__name__} at {hex(id(self))} socket={self.socket}, "
+            f"ipaddr={self.params.host}, port={self.params.port}, sslctx={self.sslctx}, "
+            f"timeout={self.params.timeout}>"
+        )
