+++ pymodbus/client/udp.py
@@ -9,7 +9,7 @@ from pymodbus.client.base import ModbusBaseClient, ModbusClientProtocol
 from pymodbus.constants import Defaults
 from pymodbus.framer import ModbusFramer
 from pymodbus.framer.socket_framer import ModbusSocketFramer
-
+from pymodbus.exceptions import ConnectionException
 
 _logger = logging.getLogger(__name__)
 
@@ -49,7 +49,7 @@ class AsyncModbusUdpClient(ModbusBaseClient):
         super().__init__(framer=framer, **kwargs)
         self.params.host = host
         self.params.port = port
-        self.source_address = source_address
+        self.params.source_address = source_address
 
         self.loop = asyncio.get_event_loop()
         self.protocol = None
@@ -159,3 +159,115 @@ class AsyncModbusUdpClient(ModbusBaseClient):
         await asyncio.sleep(self.delay_ms / 1000)
         self.delay_ms = 2 * self.delay_ms
         return await self._connect()
+
+
+class ModbusUdpClient(ModbusBaseClient):
+    """**ModbusUdpClient**.
+
+    :param host: Host IP address or host name
+    :param port: (optional) Port used for communication.
+    :param framer: (optional) Framer class.
+    :param source_address: (optional) source address of client,
+    :param kwargs: (optional) Experimental parameters
+
+    Example::
+
+        from pymodbus.client import ModbusUdpClient
+
+        async def run():
+            client = ModbusUdpClient("localhost")
+
+            client.connect()
+            ...
+            client.close()
+    """
+
+    def __init__(
+        self,
+        host: str,
+        port: int = Defaults.UdpPort,
+        framer: ModbusFramer = ModbusSocketFramer,
+        source_address: typing.Tuple[str, int] = None,
+        **kwargs: any,
+    ) -> None:
+        """Initialize Asyncio Modbus UDP Client."""
+        super().__init__(framer=framer, **kwargs)
+        self.params.host = host
+        self.params.port = port
+        self.params.source_address = source_address
+
+        self.socket = None
+
+    @property
+    def connected(self):
+        """Connect internal.
+
+        :meta private:
+        """
+        return self.connect()
+
+    def connect(self):
+        """Connect to the modbus tcp server.
+
+        :meta private:
+        """
+        if self.socket:
+            return True
+        try:
+            family = ModbusUdpClient._get_address_family(self.params.host)
+            self.socket = socket.socket(family, socket.SOCK_DGRAM)
+            self.socket.settimeout(self.params.timeout)
+        except socket.error as exc:
+            txt = f"Unable to create udp socket {exc}"
+            _logger.error(txt)
+            self.close()
+        return self.socket is not None
+
+    def close(self):
+        """Close the underlying socket connection.
+
+        :meta private:
+        """
+        self.socket = None
+
+    def send(self, request):
+        """Send data on the underlying socket.
+
+        :meta private:
+        """
+        super().send(request)
+        if not self.socket:
+            raise ConnectionException(str(self))
+        if request:
+            return self.socket.sendto(request, (self.params.host, self.params.port))
+        return 0
+
+    def recv(self, size):
+        """Read data from the underlying descriptor.
+
+        :meta private:
+        """
+        super().recv(size)
+        if not self.socket:
+            raise ConnectionException(str(self))
+        return self.socket.recvfrom(size)[0]
+
+    def is_socket_open(self):
+        """Check if socket is open.
+
+        :meta private:
+        """
+        if self.socket:
+            return True
+        return self.connect()
+
+    def __str__(self):
+        """Build a string representation of the connection."""
+        return f"ModbusUdpClient({self.params.host}:{self.params.port})"
+
+    def __repr__(self):
+        """Return string representation."""
+        return (
+            f"<{self.__class__.__name__} at {hex(id(self))} socket={self.socket}, "
+            f"ipaddr={self.params.host}, port={self.params.port}, timeout={self.params.timeout}>"
+        )
